requirements.txt
# Dependencias principales
Flask==2.3.3
Werkzeug==2.3.7
jinja2==3.1.2
python-multipart==0.0.6

# Procesamiento de documentos
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.3
openpyxl==3.1.2
psycopg2-binary==2.9.10

# Modelos de IA y procesamiento
numpy==1.26.0
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Opcionales (descomente si necesita API)
# fastapi==0.103.1
# uvicorn==0.23.2
# pydantic==2.3.0
# langchain==0.0.267 